Renamed KhronosNeutral to just Neutral tone-mapper
