+++ examples/src/examples/graphics/post-processing/controls.mjs
@@ -56,7 +56,7 @@ export function controls({ observer, ReactPCUI, React, jsx, fragment }) {
                         { v: pc.TONEMAP_HEJL, t: 'HEJL' },
                         { v: pc.TONEMAP_ACES, t: 'ACES' },
                         { v: pc.TONEMAP_ACES2, t: 'ACES2' },
-                        { v: pc.TONEMAP_KHRONOS_NEUTRAL, t: 'KHRONOS_NEUTRAL' }
+                        { v: pc.TONEMAP_NEUTRAL, t: 'NEUTRAL' }
                     ]
                 })
             )
