+++ src/framework/app-base.js
@@ -1493,7 +1493,7 @@ class AppBase extends EventHandler {
      * - {@link TONEMAP_HEJL}
      * - {@link TONEMAP_ACES}
      * - {@link TONEMAP_ACES2}
-     * - {@link TONEMAP_KHRONOS_NEUTRAL}
+     * - {@link TONEMAP_NEUTRAL}
      *
      * @param {number} settings.render.exposure - The exposure value tweaks the overall brightness
      * of the scene.
